создает абстрактный класс, выносит часть логики в контроллер
