+++ src/components/film-details.js
@@ -1,5 +1,5 @@
 import {EMOTIONS} from "../mock/film";
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component";
 
 const createCommentMarkup = (comments) => {
   return comments.map((comment) => {
@@ -157,24 +157,18 @@ const createFilmDetailsTemplate = (film) => {
   );
 };
 
-export default class FilmDetails {
+export default class FilmDetails extends AbstractComponent {
   constructor(film) {
+    super();
     this._film = film;
-    this._element = null;
   }
 
   getTemplate() {
     return createFilmDetailsTemplate(this._film);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
+  setCloseButtonClickHandler(handler) {
+    this.getElement().querySelector(`.film-details__close-btn`)
+      .addEventListener(`click`, handler);
   }
 }
