+++ src/components/films-container.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component";
 
 const createFilmsContainerTemplate = () => {
   return (
@@ -11,23 +11,8 @@ const createFilmsContainerTemplate = () => {
   );
 };
 
-export default class FilmsContainer {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FilmsContainer extends AbstractComponent {
   getTemplate() {
     return createFilmsContainerTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
