+++ src/components/profile.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component";
 
 export const createProfileTemplate = (rank) => {
   return (
@@ -9,25 +9,13 @@ export const createProfileTemplate = (rank) => {
   );
 };
 
-export default class Profile {
+export default class Profile extends AbstractComponent {
   constructor(rank) {
+    super();
     this._rank = rank;
-
-    this._element = null;
   }
 
   getTemplate() {
     return createProfileTemplate(this._rank);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
